GA Direct To: address brunotl review findings
Fix waypoint name UTF-8 truncation at the source (LK_tcsncpy -> lk::strcpy
in dlgDirectToCountdown.cpp) instead of sanitizing bytes after the fact;
close a CritSec_TaskData locking gap in SetTargetPanWaypoint; replace the
hardcoded "Direct\nTo" menu label with a proper MsgToken; drop a redundant
cast; revert the unrelated CMakeLists.txt ENABLE_OPENMP default change
(unwired to gradle, out of scope for this PR). Document why
DirectToWaypointIndex's -1 default and the overindex reuse in DrawBearing
are already correct rather than changing working code.

+++ Common/Source/InputEvents.cpp
@@ -882,7 +882,7 @@ void InputEvents::eventDirectToFromPan(const TCHAR* /*misc*/) {
     const std::lock_guard lock(CritSec_TaskData);
     WayPointList[RESWP_PANPOS].Latitude  = pan_lat;
     WayPointList[RESWP_PANPOS].Longitude = pan_lon;
-    WayPointList[RESWP_PANPOS].Altitude  = (double)th;
+    WayPointList[RESWP_PANPOS].Altitude  = th;
   }
 
   MapWindow::Event_Pan(0);  // exit pan mode before showing countdown
